test(qa): Testsuite-Audit umgesetzt -- 8 Findings gefixt, Fence gehaertet, Mock ent-zirkularisiert
/fable-review ueber die gesamte Testsuite (6 Finder-Winkel, jede
tragende Behauptung adversarial verifiziert; Verdict mit 8 bestaetigten
Findings und 6 widerlegten Claims: docs/review/2026-08-22-testsuite-
verdict.md, advisor-APPROVED). Umgesetzt:

- Golden-Fence: APRS_GOLDEN_UPDATE-Laeufe schlagen jetzt IMMER fehl
  (Regeneration nie gruen -- ein Env-Var-Leck kann den Fence nicht mehr
  stumm neutralisieren); canonical() erfasst zusaetzlich appoff/srccall/
  srclast/pathcnt (golden.txt einmalig regeneriert, Diff additiv);
  canonicalCore() prueft im Roundtrip jetzt hop/srv/track/appoff/dstcall,
  Mesh-Bit bewusst maskiert (encodeAPRS stampt es aus dem globalen bMESH,
  doc 11 §1.2); Korpus-Kapazitaet schlaegt laut fehl statt still zu
  kappen; corpus.txt und golden.txt loesen ueber denselben Pfad-Praefix.
- txring: defensive Laengen-Invariante am Engpass (len==0/>255 -> -1,
  alle heutigen Aufrufer nachweislich begrenzt) + 3 Randtests.
- Shims: BOARD_HARDWARE-ODR-Verletzung in 4 Test-TUs behoben (int statt
  uint8_t); millis-Shim-Uhr jetzt uint32_t (2^32-Wrap physisch echt);
  String::substring(from>to) an echtes Arduino-Verhalten angeglichen;
  millis-Leck in test_regex_call geschlossen.
- Mock: CONF-Tautologie (x==x) durch handgerechnetes Byte-Literal aus
  doc 11 §2.2 ersetzt (inkl. Zweierkomplement-Nachweis fuer negative
  Laengengrade); DATA-Header einmal als handgeschriebenes 36-Byte-Literal
  per Raw-Socket getestet (bricht Encoder/Decoder-Drift-Zirkularitaet);
  Registry-Expiry real getestet (ttl=0.2s); Teardown-Poll 0.5->0.1 s
  (Suite 6s -> 2.8s); README-Ueberclaims korrigiert.

Dokumentierte Nicht-Aenderungen (mit Begruendung im Verdict):
test_compress (inert, upstream), external_radio-Spiegel-Logik (upstream-
Suiten), Mock-Routing-Annahme (Wire-Shape spezifiziert, Routing nicht).

Gate: native_aprs 31/31, native 15/15, Mock 14/14; wiscore (-Werror),
heltec V3, t114, t_echo gruen; unabhaengiger Advisor (Fable) hat Diff,
Literale und Gate eigenstaendig nachgeprueft -- APPROVED.

+++ test/test_aprs_decode/test_aprs_decode.cpp
@@ -24,7 +24,7 @@ bool bDisplayInfo = false;
 bool bDisplayCont = false;
 bool bLORADEBUG = false;
 bool bMESH = true;
-uint8_t BOARD_HARDWARE = 9;   // RAK4631 (wie auf der Bench)
+int BOARD_HARDWARE = 9;   // RAK4631 (wie auf der Bench) -- int statt uint8_t: ODR-Begruendung siehe test_txring.cpp (Verdict Finding 4)
 int getMOD(void) { return 3; }
 void printAsciiBuffer(unsigned char *buf, int len) { (void)buf; (void)len; }
