+++ test/test_aprs_spec/test_aprs_spec.cpp
@@ -33,7 +33,7 @@ bool bDisplayInfo = false;
 bool bDisplayCont = false;
 bool bLORADEBUG = false;
 bool bMESH = true;
-uint8_t BOARD_HARDWARE = 9;
+int BOARD_HARDWARE = 9;   // int statt uint8_t: ODR-Begruendung siehe test_txring.cpp (Verdict Finding 4)
 int getMOD(void) { return 3; }
 void printAsciiBuffer(unsigned char *buf, int len) { (void)buf; (void)len; }
